Create page file
