+++ lib/index.ts
@@ -1,4 +1,5 @@
 export * from './module';
+export * from './page';
 export * from './pib';
 export * from './provider';
 export * from './storage';
