+++ lib/module.ts
@@ -1,4 +1,5 @@
 import { Provider } from './provider';
+import { Page }     from './page';
 
 /**
  * A Pib module.
@@ -10,7 +11,7 @@ export abstract class Module {
 
     protected _path?: string;
 
-    protected _pages: Module.Page[];
+    protected _pages: Page[];
 
     protected constructor(name: string, slug?: string, path?: string) {
         this._name = name;
@@ -24,7 +25,7 @@ export abstract class Module {
         return this._name;
     }
 
-    public add(page: Module.Page): void {
+    public add(page: Page): void {
         this._pages = this._pages.concat(page);
     }
 
@@ -95,49 +96,6 @@ export namespace Module {
 
             public get nullable(): boolean {
                 return this.#nullable;
-            }
-        }
-    }
-
-    /**
-     * A module page.
-     * This is a page for a module. This should be extended whenever creating any sort of page. The page will specify
-     * a layout and will be specified different settings and variables when the page is requested so that it can
-     * dynamically display different content depending on the current context of the request.
-     */
-    export abstract class Page {
-        protected constructor() {
-
-        }
-
-        public abstract load(provider: Provider): void;
-    }
-
-    /**
-     * The page namespace.
-     * This namespace defines all of the different things which are useful for a page.
-     */
-    export namespace Page {
-        /**
-         * A page layout.
-         * This is the top-level overview of a page's content. It is what should be defining how a page should be
-         * laid out.
-         */
-        export abstract class Layout {
-            protected constructor() {
-
-            }
-        }
-
-        /**
-         * A Pib component.
-         * A component is an element on a page. Every page is made up of some number of components. Components could be
-         * anything from a text box to a simple container. Components should be made to be as flexible and reusable
-         * as possible so that only a few components need to be made.
-         */
-        export abstract class Component {
-            protected constructor() {
-
             }
         }
     }
